+++ Source/AdMobAdapter.swift
@@ -29,7 +29,7 @@ final class AdMobAdapter: PartnerAdapter {
     /// The version of the adapter.
     /// It should have either 5 or 6 digits separated by periods, where the first digit is Chartboost Mediation SDK's major version, the last digit is the adapter's build version, and intermediate digits are the partner SDK's version.
     /// Format: `<Chartboost Mediation major version>.<Partner major version>.<Partner minor version>.<Partner patch version>.<Partner build version>.<Adapter build version>` where `.<Partner build version>` is optional.
-    let adapterVersion = "4.9.14.0.0"
+    let adapterVersion = "4.9.14.0.1"
     
     /// The partner's unique identifier.
     let partnerIdentifier = "admob"
@@ -139,7 +139,9 @@ final class AdMobAdapter: PartnerAdapter {
             return AdMobAdapterInterstitialAd(adapter: self, request: request, delegate: delegate, extras: sharedExtras)
         case .rewarded:
             return AdMobAdapterRewardedAd(adapter: self, request: request, delegate: delegate, extras: sharedExtras)
-        @unknown default:
+        case .rewardedInterstitial:
+            return AdMobAdapterRewardedInterstitialAd(adapter: self, request: request, delegate: delegate, extras: sharedExtras)
+        default:
             throw error(.loadFailureUnsupportedAdFormat)
         }
     }