+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Mediation SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
+### 4.9.14.0.1
+- Added support for the Rewarded Interstitial ad format.
+- This version of the adapters has been certified with Google-Mobile-Ads-SDK 9.14.0.
+
 ### 4.9.14.0.0
 - This version of the adapters has been certified with Google-Mobile-Ads-SDK 9.14.0.